requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
coverage==4.5.3
Flask==1.0.2
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
pluggy==0.9.0
psycopg2==2.8.2
py==1.8.0
pycodestyle==2.5.0
pytest==4.4.1
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.25.3
Werkzeug==0.15.2